optimization of views

+++ src/main/java/org/ohdsi/webapi/cohortcharacterization/repository/AnalysisGenerationInfoEntityRepository.java
@@ -3,5 +3,5 @@ package org.ohdsi.webapi.cohortcharacterization.repository;
 import com.cosium.spring.data.jpa.entity.graph.repository.EntityGraphJpaRepository;
 import org.ohdsi.webapi.common.generation.AnalysisGenerationInfoEntity;
 
-public interface AnalysisGenerationInfoEntityRepository extends EntityGraphJpaRepository<AnalysisGenerationInfoEntity, Integer> {
+public interface AnalysisGenerationInfoEntityRepository extends EntityGraphJpaRepository<AnalysisGenerationInfoEntity, Long> {
 }
